requirements.txt
ninja
torch==1.13.0
torchvision==0.14.0
opencv-python
numpy
pillow
imageio
pyrender
matplotlib
einops
scipy
typing
trimesh
tqdm
timm
facenet_pytorch
face_alignment
segmentation_models_pytorch